+++ site/src/versions.ts
@@ -2,6 +2,7 @@ declare const __BEWITCH_VERSION__: string
 
 export const versions = [
   { label: `v${__BEWITCH_VERSION__}-dev`, path: '/docs' },
+  { label: 'v0.4.0', path: '/docs/v0.4.0' },
   { label: 'v0.3.1', path: '/docs/v0.3.1' },
   { label: 'v0.3.0', path: '/docs/v0.3.0' },
   { label: 'v0.2.0', path: '/docs/v0.2.0' },